+++ examples/nextjs/pages/_app.tsx
@@ -1,12 +1,17 @@
-import '../styles/globals.css';
-import type { AppProps } from 'next/app';
-import { Wallet01 } from '@wallet01/react';
-import { CoinbaseConnector, InjectedConnector } from '@wallet01/evm';
-import { KeplrConnector } from '@wallet01/cosmos';
-import { PhantomConnector, SolflareConnector } from '@wallet01/solana';
-import { FilecoinConnector } from '../lib/filecoin';
-import Layout from '../components/layout';
-import { WalletconnectConnector } from '@wallet01/evm';
+import "../styles/globals.css";
+import type { AppProps } from "next/app";
+
+import { Wallet01 } from "@wallet01/react";
+import {
+  CoinbaseConnector,
+  InjectedConnector,
+  WalletconnectConnector,
+} from "@wallet01/evm";
+import { KeplrConnector } from "@wallet01/cosmos";
+import { PhantomConnector, SolflareConnector } from "@wallet01/solana";
+
+import Layout from "../components/layout";
+import { FilecoinConnector } from "../lib/filecoin";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (