+++ packages/core/src/index.ts
@@ -1,15 +1,15 @@
 // import { Client, createClient } from './client';
-import Client from './client/client';
-import { useStore } from './store/rootStore';
+import Client from "./client/client";
+import { useStore } from "./store/rootStore";
 
 import {
   type ClientConfig,
   type ConnectedData,
   BaseConnector,
   TProvider,
-} from './types';
+} from "./types";
 
-import setLastUsedConnector from './utils/util';
+import setLastUsedConnector from "./utils/util";
 
 export {
   Client,