+++ packages/cosmos/src/connectors/keplr.ts
@@ -1,39 +1,35 @@
-import { Window as KeplrWindow } from '@keplr-wallet/types';
-import { KeplrProvider } from '../providers/keplrProvider';
-import { BaseConnector, setLastUsedConnector } from '@wallet01/core';
-import emitter from '../utils/emiter';
+import { Window as KeplrWindow } from "@keplr-wallet/types";
+import { KeplrProvider } from "../providers/keplrProvider";
+import { BaseConnector, setLastUsedConnector } from "@wallet01/core";
+import emitter from "../utils/emiter";
 
 declare const window: KeplrWindow;
 
 export class KeplrConnector extends BaseConnector<KeplrProvider> {
   provider!: KeplrProvider;
-  chain: string;
-  name: string;
 
-  constructor(chain: string = 'secret-4') {
-    super(chain);
-    this.chain = chain;
-    this.name = 'Keplr';
+  constructor(chain: string = "secret-4") {
+    super(chain, "keplr");
   }
 
   async getProvider(): Promise<KeplrProvider> {
-    if (typeof window !== 'undefined' && window.keplr) {
+    if (typeof window !== "undefined" && window.keplr) {
       this.provider = window.keplr;
       return this.provider;
     } else {
-      throw new Error('Wallet Not Installed!');
+      throw new Error("Wallet Not Installed!");
     }
   }
 
   async getAccount(): Promise<string[]> {
-    if (!this.provider) throw new Error('Provider undefined');
+    if (!this.provider) throw new Error("Provider undefined");
     try {
       await this.provider.enable(this.chain);
       const { bech32Address } = await this.provider.getKey(this.chain);
       return [bech32Address];
     } catch (error) {
       console.error(error);
-      throw new Error('Error in getting accounts');
+      throw new Error("Error in getting accounts");
     }
   }
 
@@ -42,7 +38,7 @@ export class KeplrConnector extends BaseConnector<KeplrProvider> {
   }
 
   async switchChain(chainId: string): Promise<void> {
-    if (!this.provider) throw new Error('Provider undefined');
+    if (!this.provider) throw new Error("Provider undefined");
     try {
       await this.provider.enable(chainId);
       this.chain = chainId;
@@ -52,10 +48,10 @@ export class KeplrConnector extends BaseConnector<KeplrProvider> {
     }
   }
 
-  async connect({ chainId = 'secret-4' }): Promise<void> {
+  async connect({ chainId = "secret-4" }): Promise<void> {
     try {
       const provider = await this.getProvider();
-      if (!provider) throw new Error('Keplr not installed');
+      if (!provider) throw new Error("Keplr not installed");
 
       await this.provider.enable(chainId);
       setLastUsedConnector(this.name);
@@ -66,8 +62,8 @@ export class KeplrConnector extends BaseConnector<KeplrProvider> {
   }
 
   async disconnect(): Promise<void> {
-    this.chain = '';
-    emitter.emit('disconnected');
+    this.chain = "";
+    emitter.emit("disconnected");
   }
 
   async resolveDid(_address: string): Promise<string | null> {
@@ -76,13 +72,15 @@ export class KeplrConnector extends BaseConnector<KeplrProvider> {
   }
 
   async signMessage(message: string): Promise<string> {
-    if (!this.provider) throw new Error('Provider Undefined');
+    if (!this.provider) throw new Error("Provider Undefined");
     try {
+      const address = (await this.getAccount())[0];
+
+      if (!address) throw new Error("No address found");
+
       const { signature } = await this.provider.signArbitrary(
         this.chain,
-        (
-          await this.getAccount()
-        )[0],
+        address,
         message
       );
       return signature;
@@ -93,14 +91,14 @@ export class KeplrConnector extends BaseConnector<KeplrProvider> {
   }
 
   protected onAccountsChanged(): void {
-    console.log('Account Changed');
+    console.log("Account Changed");
   }
 
   protected onChainChanged(_chain: string | number): void {
-    console.log('Chain Changed');
+    console.log("Chain Changed");
   }
 
   protected onDisconnect(): void {
-    console.log('Wallet disconnected');
+    console.log("Wallet disconnected");
   }
 }