+++ examples/vite-react/CHANGELOG.md
@@ -1,5 +1,15 @@
 # vite-react
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @wallet01/cosmos@0.5.0
+  - @wallet01/evm@0.7.0
+  - @wallet01/react@0.6.0
+  - @wallet01/solana@0.5.0
+
 ## 0.0.15
 
 ### Patch Changes